Major refactoring in preparation for mouse-over highlight of nets.
Drawing of wires, junctions, and labels are now pre-computed in typed
arrays, storing line segments, dots, and label positions.

Actual drawing then consists just of iterating over the arrays,
outputting drawing commands.

This in preparation for being able to find and highligt the supernet
that is under the mouse pointer.