Auto-zoom in on a net selected from the listbox.
The idea is to make the newly selected net visible if it is not so
already at the current zoom, so that the selection can be immediately
seen to take effect.

The heuristics to find a "good" zoom level works as follows:

1. If the supernet is active in a tile that is already visible at the
current zoom level, do nothing.

2. Try to set the zoom level to show all tiles in which the selected
supernet is active. If wires are still visible at this zoom-level, use
it.

3. Otherwise, pick an arbitrary tile in which the supernet is active,
and zoom in centered on that tile, with a zoom level that is just
enough to have wires still visible.