Use inline javascript in standalone mode
In standalone mode, copy the various referenced *.js files directly into the
generated .html file, rather than referencing them with file:/// urls.

This helps avoid version problems when ice40_viewer is upgraded. With the
file:/// urls, an old output .html might reference new, replaced .js files,
which is likely to cause problems.

The chipdb-8k.txt.js is still left as file:/// reference. This file is 40
MB, so not nice to include in every output file, and it is not likely to
change in an incompatible way (or at all).