Implement support for ICE40 HX1K devices.
The chipdb is now included as verbatim strings in a loaded .js file.
One for 1k and one for 8k devices.

The chipdb is selected early during .asc parsing, when the header is
seen that tells which of the two devices the .asc is for.

Also add an 1k example (the `icestick' example from project Icestorm).